Cascaded process deletes to activity-assignments

+++ src/main/scala/com/buildwhiz/baf2/ProcessApi.scala
@@ -29,6 +29,10 @@ object ProcessApi {
     if (isActive(process))
       throw new IllegalArgumentException(s"Process '${process.name[String]}' is still active")
 
+    val assignmentDeleteResult = BWMongoDB3.activity_assignments.deleteOne(Map("process_id" -> processOid))
+    if (assignmentDeleteResult.getDeletedCount == 0)
+      throw new IllegalArgumentException(s"MongoDB error: $assignmentDeleteResult")
+
     val processDeleteResult = BWMongoDB3.processes.deleteOne(Map("_id" -> processOid))
     if (processDeleteResult.getDeletedCount == 0)
       throw new IllegalArgumentException(s"MongoDB error: $processDeleteResult")
